requirements.txt
absl-py==1.4.0
appdirs==1.4.4
argon2-cffi==21.1.0
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.1.0
blinker==1.4
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
Cython==3.0.7
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dm-tree==0.1.8
entrypoints==0.4
etils==1.6.0
exceptiongroup==1.2.0
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.47.0
fs==2.4.12
fsspec==2023.12.2
gast==0.5.4
gin-config==0.5.0
google-api-core==2.15.0
google-api-python-client==2.112.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h5py==3.10.0
html5lib==1.1
httplib2==0.22.0
idna==3.6
imageio==2.34.0
immutabledict==4.1.0
importlib-metadata==4.6.4
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.19.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
jsonschema==3.2.0
jupyter-console==6.4.0
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyterlab-pygments==0.1.2
kaggle==1.6.0
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lxml==5.0.1
lz4==3.1.3+dfsg
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==0.0.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.8
netifaces==0.11.0
notebook==6.4.8
numpy==1.26.4
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
opencv-python==4.5.5.64
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
portalocker==2.8.2
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.1.1
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-slugify==8.0.1
pythran==0.10.0
pytz==2023.3.post1
pyusb==1.2.1.post1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
remotezip==0.12.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.4.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
SecretStorage==3.3.1
Send2Trash==1.8.1b0
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
soupsieve==2.3.1
stack-data==0.6.3
sympy==1.9
systemd-python==234
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-datasets==4.9.4
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
tensorflow-model-optimization==0.7.5
tensorflow-text==2.15.0
termcolor==2.4.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.15.0
tf-slim==1.1.0
threadpoolctl==3.2.0
toml==0.10.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==2.1.0
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websockets==9.1
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==2.0.0
wrapt==1.14.1
zipp==3.17.0
